moved rari project into projects dir
