+++ projects/rari/abi.json
@@ -186,6 +186,27 @@
     ],
     "stateMutability": "nonpayable",
     "type": "function"
-  }
+  },
+  "balanceOf":{
+		"constant": true,
+		"inputs": [
+			{
+				"internalType": "address",
+				"name": "account",
+				"type": "address"
+			}
+		],
+		"name": "balanceOf",
+		"outputs": [
+			{
+				"internalType": "uint256",
+				"name": "",
+				"type": "uint256"
+			}
+		],
+		"payable": false,
+		"stateMutability": "view",
+		"type": "function"
+	}
 
 }
